+++ scripts/dialog_handling.sh
@@ -335,7 +335,7 @@ menu_parse() {
                 alt_command "$label" "$key" "$cmd" "$keep_cmd"
             else
                 mnu_command "$label" "$key" "$cmd" "$keep_cmd"
-                $b_do_show_cmds && show_cmd "$TMUX_BIN $cmd"
+                $b_do_show_cmds && sc_show_cmd "$TMUX_BIN $cmd"
             fi
             ;;
 
@@ -364,19 +364,13 @@ menu_parse() {
             # first extract the variables, then  if it shouldn't be used move on
             ! tmux_vers_check "$min_vers" && continue
 
-            #
-            #  Expand relative PATH at one spot, before calling the
-            #  various implementations
-            #
-            echo "$cmd" | grep -vq / && cmd="$d_scripts/$cmd"
-
             [ -n "$menu_debug" ] && debug_print "key[$key] label[$label] command[$cmd]"
 
             if $cfg_use_whiptail; then
                 alt_external_cmd "$label" "$key" "$cmd"
             else
                 mnu_external_cmd "$label" "$key" "$cmd"
-                $b_do_show_cmds && [ "$key" != "!" ] && show_cmd "$cmd"
+                $b_do_show_cmds && [ "$key" != "!" ] && sc_show_cmd "$cmd"
             fi
             ;;
 
@@ -874,7 +868,10 @@ ensure_menu_fits_on_screen() {
 
 clear_prep_disp_status() {
     safe_now
-    log_it "Preparing Display Commands took: $(echo "$t_now - $t_show_cmds" | bc)s"
+    display_command_label
+    log_it "$(relative_path "$0") - Preparing $_lbl took: $(
+        echo "$t_now - $t_show_cmds" | bc
+    )s"
 
     if tmux_vers_check 3.2; then
         tmux_error_handler display-message -d 1 ""
@@ -932,19 +929,22 @@ prepare_show_commands() {
     # action item
     # log_it "prepare_show_commands()"
 
+    $all_helpers_sourced || source_all_helpers "prepare_show_commands()"
+
     # Do this before the timer is started, otherwise the first usage of show commands
     # will always be slower
     $all_helpers_sourced || source_all_helpers "prepare_show_commands"
     [ ! -f "$f_cached_tmux_key_binds" ] && {
         log_it "Creating: $f_cached_tmux_key_binds"
+        # Filtering out all key binds displaying a menu, since they won't be relevant
         $TMUX_BIN list-keys | grep -iv display-menu >"$f_cached_tmux_key_binds"
     }
 
     safe_now t_show_cmds
     cfg_use_cache=false
     b_do_show_cmds=true
-    tmux_error_handler display-message "Preparing Display Commands ..."
-
+    display_command_label
+    tmux_error_handler display-message "Preparing $_lbl ..."
     # shellcheck source=scripts/show_cmd.sh
     . "$D_TM_BASE_PATH"/scripts/show_cmd.sh
 }
@@ -954,23 +954,12 @@ display_commands_toggle() {
     # log_it "display_commands_toggle($menu_part)"
     [ -z "$menu_part" ] && error_msg "add_display_commands() - called with no param"
 
-    case "$TMUX_MENUS_SHOW_CMDS" in
-    "1")
-        _lbl="Display key binds"
-        _i=2
-        ;;
-    "2")
-        _lbl="Hide key binds"
-        _i=0
-        ;;
-    *)
-        _lbl="Display Commands"
-        _i=1
-        ;;
-    esac
+    # In case we got here via dynamic_content()
+    $all_helpers_sourced || source_all_helpers "display_commands_toggle()"
 
+    display_command_label
     set -- \
-        0.0 E ! "$_lbl" "TMUX_MENUS_SHOW_CMDS=$_i $0"
+        0.0 E ! "$_lbl_next" "TMUX_MENUS_SHOW_CMDS='$_idx_next' $0"
 
     menu_generate_part "$menu_part" "$@"
 }